+++ pybamm/models/submodels/particle/fast_many_particles.py
@@ -4,10 +4,10 @@
 #
 import pybamm
 
-from .base_fast_particle import BaseModel
+from .base_particle import BaseParticle
 
 
-class ManyParticles(BaseModel):
+class FastManyParticles(BaseParticle):
     """Base class for molar conservation in many particles with
     uniform concentration in r (i.e. infinitely fast diffusion within particles).
 
@@ -19,7 +19,7 @@ class ManyParticles(BaseModel):
         The domain of the model either 'Negative' or 'Positive'
 
 
-    **Extends:** :class:`pybamm.particle.fast.BaseModel`
+    **Extends:** :class:`pybamm.particle.BaseParticle`
     """
 
     def __init__(self, param, domain):
@@ -33,7 +33,7 @@ class ManyParticles(BaseModel):
             c_s = pybamm.PrimaryBroadcast(c_s_surf, ["negative particle"])
             c_s_xav = pybamm.x_average(c_s)
 
-            N_s = pybamm.FullBroadcast(
+            N_s = pybamm.FullBroadcastToEdges(
                 0,
                 ["negative particle"],
                 auxiliary_domains={
@@ -41,14 +41,14 @@ class ManyParticles(BaseModel):
                     "tertiary": "current collector",
                 },
             )
-            N_s_xav = pybamm.x_average(N_s)
+            N_s_xav = pybamm.FullBroadcast(0, "negative electrode", "current collector")
 
         elif self.domain == "Positive":
             c_s_surf = pybamm.standard_variables.c_s_p_surf
             c_s = pybamm.PrimaryBroadcast(c_s_surf, ["positive particle"])
             c_s_xav = pybamm.x_average(c_s)
 
-            N_s = pybamm.FullBroadcast(
+            N_s = pybamm.FullBroadcastToEdges(
                 0,
                 ["positive particle"],
                 auxiliary_domains={
@@ -56,23 +56,24 @@ class ManyParticles(BaseModel):
                     "tertiary": "current collector",
                 },
             )
-            N_s_xav = pybamm.x_average(N_s)
+            N_s_xav = pybamm.FullBroadcast(0, "positive electrode", "current collector")
 
         variables = self._get_standard_concentration_variables(c_s, c_s_xav)
         variables.update(self._get_standard_flux_variables(N_s, N_s_xav))
 
         return variables
 
-    def _unpack(self, variables):
+    def set_rhs(self, variables):
         c_s_surf = variables[self.domain + " particle surface concentration"]
-        N_s = variables[self.domain + " particle flux"]
         j = variables[self.domain + " electrode interfacial current density"]
+        if self.domain == "Negative":
+            self.rhs = {c_s_surf: -3 * j / self.param.a_n}
 
-        return c_s_surf, N_s, j
+        elif self.domain == "Positive":
+            self.rhs = {c_s_surf: -3 * j / self.param.a_p / self.param.gamma_p}
 
     def set_initial_conditions(self, variables):
-        c, _, _ = self._unpack(variables)
-
+        c_s_surf = variables[self.domain + " particle surface concentration"]
         if self.domain == "Negative":
             x_n = pybamm.standard_spatial_vars.x_n
             c_init = self.param.c_n_init(x_n)
@@ -81,4 +82,4 @@ class ManyParticles(BaseModel):
             x_p = pybamm.standard_spatial_vars.x_p
             c_init = self.param.c_p_init(x_p)
 
-        self.initial_conditions = {c: c_init}
+        self.initial_conditions = {c_s_surf: c_init}