+++ tests/unit/test_models/test_submodels/test_particle/test_fickian_many_particles.py
@@ -1,5 +1,5 @@
 #
-# Test many fast particles
+# Test many fickian particles
 #
 
 import pybamm
@@ -18,14 +18,20 @@ class TestManyParticles(unittest.TestCase):
             pybamm.Scalar(0), "positive electrode", {"secondary": "current collector"}
         )
 
-        variables = {"Negative electrode interfacial current density": a_n}
+        variables = {
+            "Negative electrode interfacial current density": a_n,
+            "Negative electrode temperature": a_n,
+        }
 
-        submodel = pybamm.particle.fast.ManyParticles(param, "Negative")
+        submodel = pybamm.particle.FickianManyParticles(param, "Negative")
         std_tests = tests.StandardSubModelTests(submodel, variables)
         std_tests.test_all()
 
-        variables = {"Positive electrode interfacial current density": a_p}
-        submodel = pybamm.particle.fast.ManyParticles(param, "Positive")
+        variables = {
+            "Positive electrode interfacial current density": a_p,
+            "Positive electrode temperature": a_p,
+        }
+        submodel = pybamm.particle.FickianManyParticles(param, "Positive")
         std_tests = tests.StandardSubModelTests(submodel, variables)
         std_tests.test_all()